Move profile files to the profile folder (Profile)
